==> 2-QC.ipynb <==
# Quality Control

import os, sys
from pathlib import Path

import pandas as pd
import numpy as np

# import utils.py
currentdir = os.path.dirname(os.path.realpath(__name__))
parentdir = os.path.dirname(currentdir)
sys.path.append(parentdir)

from utils import *

config_file = "config.toml"
config = parse_config(config_file)
root_dir = Path(config['root_dir'])

### Input

raw_fpath_file = root_dir / 'raw_fpath.csv'
raw_fpath_file

### Onput

trimmed_fpath_file = root_dir / 'trimmed_fpath.csv'
trimmed_fpath_file

## 1. Load file path

fpath_df = pd.read_csv(raw_fpath_file)
fpath_df

## 2. Encode stat

### Bases

def get_seqkit_stat(fastq):
    cmd = f"seqkit stat {fastq}"
    output = subprocs(cmd)
    data = [line.split() for line in output.stdout.decode().strip().split('\n')]
    if len(data[1]) == 6:
        v = [data[1][0], data[1][1], data[1][2], 0, 0, 0, 0.0, 0]
        data[1] = v
    stat_ser = pd.Series(data[1], index=data[0])
    for col in ['num_seqs', 'sum_len', 'min_len', 'max_len']:
        stat_ser[col] = int(stat_ser[col].replace(',',''))
    for col in ['avg_len']:
        stat_ser[col] = float(stat_ser[col].replace(',',''))
    return stat_ser

def get_paired_fastq_stat(r1_fastq, r2_fastq):
    ser1 = get_seqkit_stat(r1_fastq)
    ser2 = get_seqkit_stat(r2_fastq)
    
    num_seqs = ser1['num_seqs'] + ser2['num_seqs']
    sum_len = ser1['sum_len'] + ser2['sum_len']
    
    return num_seqs, sum_len, sum_len / num_seqs

out, elp = execute_function_pool_args(
    get_paired_fastq_stat,
    fpath_df[['raw_r1', 'raw_r2']].values,
    4,
)

elp

base_stat_df = pd.DataFrame(out, columns = ['num_seqs', 'sum_len', 'avg_len'])
base_stat_df

!seqkit stat data/raw/SRR6046075_1.fastq.gz data/raw/SRR6046075_2.fastq.gz

!seqkit stat -a data/raw/SRR6046075_1.fastq.gz data/raw/SRR6046075_2.fastq.gz

### Error rate

from Bio import SeqIO
import math
import gzip

def cal_error_rate(r1_fastq, r2_fastq):
    phred_qual_counts = [0 for _ in range(43)]
    for fastq in (r1_fastq, r2_fastq):
        with gzip.open(fastq, 'rt') as hdl:
            for record in SeqIO.parse(hdl, "fastq"):
                for q in record.letter_annotations['phred_quality']:
                    phred_qual_counts[q] += 1
    
    err_count = 0
    for idx, count in enumerate(phred_qual_counts):
        err_count += math.pow(10, idx/(-10)) * count
    
    return err_count / sum(phred_qual_counts)

out, elp = execute_function_pool_args(
    cal_error_rate,
    fpath_df[['raw_r1', 'raw_r2']].values,
    4,
)

elp

error_rates = pd.Series(out)
error_rates.name = 'error_rate'
error_rates

## 3. FastQC

### Download

fastqc_bin = root_dir / 'bin' / 'fastqc'
fastqc_bin.parent.mkdir(exist_ok=True)
fastqc_bin

def download_fastqc_to(bin_path):
    zip_file = bin_path.with_suffix('.zip')
    if not zip_file.exists():
        url = "https://www.bioinformatics.babraham.ac.uk/projects/fastqc/fastqc_v0.12.1.zip"
        cmd = f"wget -O {zip_file} {url}"
        subprocs(cmd)
    org_bin = (zip_file.parent / 'FastQC' / 'fastqc').absolute()
    if not org_bin.exists():
        subprocs(f"unzip -d {zip_file.parent} {zip_file}")
    if not bin_path.exists():
        bin_path.symlink_to(org_bin)

download_fastqc_to(fastqc_bin)

### Run

fastqc_out_dir = root_dir / 'fastqc_reports'
fastqc_out_dir.mkdir(exist_ok=True)
fastqc_out_dir

def run_fastqc(fastq):
    out_html_file = fastqc_out_dir / f"{Path(fastq).name.split('.')[0]}_fastqc.html"
    if not out_html_file.exists():
        cmd = f"{fastqc_bin} -o {fastqc_out_dir} {fastq}"
        subprocs(cmd)
    return out_html_file

out, elp = execute_function_pool(
    run_fastqc,
    fpath_df[['raw_r1', 'raw_r2']].values.reshape(-1),
    4,
)

# elp
107.81914901733398

fastqc_report_df = pd.DataFrame(np.array(out).reshape(-1, 2))
fastqc_report_df.columns = ['fastqc_report_r1', 'fastqc_report_r2']
fastqc_report_df

## 4. Trimming

trimmed_dir = root_dir / 'trimmed'
trimmed_dir.mkdir(exist_ok=True)
trimmed_dir

def run_trimming(r1_fastq, r2_fastq):
    r1_fastq = Path(r1_fastq)
    r2_fastq = Path(r2_fastq)
    
    trimmed_r1 = trimmed_dir / (r1_fastq.name.split('.')[0]+'_trimmed.fastq.gz')
    unpaired_r1 = trimmed_dir / (r1_fastq.name.split('.')[0]+'_unpaired.fastq.gz')
    trimmed_r2 = trimmed_dir / (r2_fastq.name.split('.')[0]+'_trimmed.fastq.gz')
    unpaired_r2 = trimmed_dir / (r2_fastq.name.split('.')[0]+'_unpaired.fastq.gz')
    log_file = trimmed_dir / (r1_fastq.name.split('_1')[0]+'.log')
    
    if not (
        trimmed_r1.exists() &
        unpaired_r1.exists() &
        trimmed_r2.exists() &
        unpaired_r2.exists() &
        log_file.exists()
    ):
        cmd = f"trimmomatic PE {r1_fastq} {r2_fastq} \
            {trimmed_r1} {unpaired_r1} \
            {trimmed_r2} {unpaired_r2} \
            TRAILING:3 SLIDINGWINDOW:4:15 MINLEN:25 -phred33 2> {log_file}"
        subprocs(cmd)
    
    return trimmed_r1, unpaired_r1, trimmed_r2, unpaired_r2, log_file

out, elp = execute_function_pool_args(
    run_trimming,
    fpath_df[['raw_r1', 'raw_r2']].values,
    4,
)

# elp
354.85028409957886

trimmed_fpath_df = pd.DataFrame(out, columns=['trimmed_r1', 'unpaired_r1', 'trimmed_r2', 'unpaired_r2', 'trimmomatic_log'])
trimmed_fpath_df

## 5. Merge and save

new_fpath_df = pd.concat([
    fpath_df,
    base_stat_df,
    error_rates,
    fastqc_report_df,
    trimmed_fpath_df,
], axis=1)
new_fpath_df

new_fpath_df.to_csv(trimmed_fpath_file, index=False)
