==> 3-Alignment.ipynb <==
# Alignment

import os, sys
from pathlib import Path

import pandas as pd
import numpy as np

# import utils.py
currentdir = os.path.dirname(os.path.realpath(__name__))
parentdir = os.path.dirname(currentdir)
sys.path.append(parentdir)

from utils import *

config_file = "config.toml"
config = parse_config(config_file)
root_dir = Path(config['root_dir'])

### Input

trimmed_fpath_file = root_dir / 'trimmed_fpath.csv'
trimmed_fpath_file

### Output

ref_file = root_dir / 'reference' / 'ref.fa'
ref_file.parent.mkdir(exist_ok=True)
ref_file

aligned_fpath_file = root_dir / 'aligned_fpath.csv'
aligned_fpath_file

## 1. Load file path

fpath_df = pd.read_csv(trimmed_fpath_file)
fpath_df

## 2. Get reference
 - <i>H37Rv</i>: https://www.ncbi.nlm.nih.gov/nuccore/NC_000962.3/

from Bio import Entrez
from Bio import SeqIO

def download_ref_file(
    accession_number,
    ref_file,
):
    if not ref_file.exists():
        Entrez.email = "your.email@example.com"

        # Fetch
        with Entrez.efetch(db="nucleotide", id=accession_number, rettype="fasta", retmode="text") as handle:
            record = SeqIO.read(handle, "fasta")

        # Save
        output_file = f"{accession_number}.fasta"
        SeqIO.write(record, ref_file, "fasta")

download_ref_file(
    'NC_000962.3',
    ref_file,
)

print(subprocs(f"seqkit stat {ref_file}").stdout.decode())

## 3. Indexing

cmd = f"bwa index {ref_file}"

out = subprocs(cmd)

out.returncode

## 4. Perform Alignment

sorted_bam_dir = root_dir / 'sorted_bam'
sorted_bam_dir.mkdir(exist_ok=True)
sorted_bam_dir

def align_with_bwa(running_id, trimmed_r1, trimmed_r2):
    bam_file = sorted_bam_dir / f"{running_id}.bam"

    if not bam_file.exists():
        cmd = f'bwa mem -aM {ref_file} {trimmed_r1} {trimmed_r2} | samtools sort > {bam_file}'
        subprocs(cmd)
    
    return bam_file

out, elp = execute_function_pool_args(
    align_with_bwa,
    fpath_df[['running_id', 'trimmed_r1', 'trimmed_r2']].values,
    4,
)

# elp
495.32800006866455

sorted_bams = pd.Series(out)
sorted_bams.name = 'sorted_bam'
sorted_bams

## 5. Merge and save

new_fpath_df = pd.concat([fpath_df, sorted_bams], axis=1)
new_fpath_df

new_fpath_df.to_csv(aligned_fpath_file, index=False)
