==> 4-CallVariants.ipynb <==
# Call variants

import os, sys
from pathlib import Path

import pandas as pd
import numpy as np

# import utils.py
currentdir = os.path.dirname(os.path.realpath(__name__))
parentdir = os.path.dirname(currentdir)
sys.path.append(parentdir)

from utils import *

config_file = "config.toml"
config = parse_config(config_file)
root_dir = Path(config['root_dir'])

### Input

ref_file = root_dir / 'reference' / 'ref.fa'
ref_file

aligned_fpath_file = root_dir / 'aligned_fpath.csv'
aligned_fpath_file

### Output

out_vcf_file = root_dir / 'out.vcf'
out_vcf_file

## 1. Load file path

fpath_df = pd.read_csv(aligned_fpath_file)
fpath_df

## 2. Pepare chunked call

chunked_call_dir = root_dir / 'chunked_call'
chunked_call_dir.mkdir(exist_ok=True)
chunked_call_dir

### (1) Indexing bam files

bam_file_list = fpath_df['sorted_bam'].values

def indexing_bam(bam_file):
    fai_file = Path(f"{bam_file}.bai")
    if not fai_file.exists():
        subprocs(f"samtools index {bam_file}")

_, elp = execute_function_pool(
    indexing_bam,
    bam_file_list,
    4,
)

# elp
10.196346998214722

### (2) write bam list to file

bam_list_file = chunked_call_dir / 'bam_list'
bam_list_file

with open(bam_list_file, 'wt') as f:
    for bam_file in bam_file_list:
        f.write(f"{bam_file}\n")

### (3) divide chunk

from Bio import SeqIO

seq_df = pd.DataFrame([
    [record.id, len(record.seq)]
    for record
    in SeqIO.parse(ref_file, 'fasta')
], columns=['contig', 'len'])

seq_df.head()

def divide_to_chunk(row, chunk_size=20000):
    data = []
    
    contig = row['contig']
    length = row['len']
    
    start_pos = 1
    end_pos = -1
    
    while end_pos != length:
        end_pos = start_pos + chunk_size - 1
        if end_pos > length:
            end_pos = length
        
        data.append([contig, start_pos, end_pos])
        start_pos = end_pos + 1
    
    df = pd.DataFrame(data, columns = ['contig', 'start', 'end'])
    return df

chunk_df = pd.concat(seq_df.apply(divide_to_chunk, axis=1).values, axis=0, ignore_index=True)
chunk_df

### (4) encode cmd (mplieup & call)

def get_chunked_call_df(
    chunk_df,
    call_res_dir = chunked_call_dir,
    ref_fa_file = ref_file,
    bam_list_file = bam_list_file,
):
    # https://samtools.github.io/bcftools/bcftools.html#common_options
    def get_region_df_while_writing(
        call_res_dir,
        chrom,
        total_len,
        chunk_size,
    ):
        data = []
        region_file_list = []
        for i in range((total_len // chunk_size) + 1):
            begin = (chunk_size * i) + 1
            if i == total_len // chunk_size:
                end = total_len
            else:
                end = begin + chunk_size - 1
            
            # write to file
            region_file = call_res_dir / f"{chrom}_{begin}_{end}"
            with open(region_file, 'w') as f:
                f.write(f"{chrom}\t{begin}\t{end}")
            region_file_list.append(region_file)
            
            data.append([i, chrom, begin, end])
            
        df = pd.DataFrame(data)
        df.columns = ['index', 'chrom', 'begin', 'end']
        df['region_file'] = region_file_list
        return df
    def write_region_file(row):
        chrom = row['contig']
        begin = row['start']
        end = row['end']
        
        region_file = call_res_dir / f"{chrom}_{begin}_{end}.region"
        with open(region_file, 'w') as f:
            f.write(f"{chrom}\t{begin}\t{end}")
        
        return region_file
    
    def encode_cmd(row):
        out_file = row['vcf_file']
        cmd = f"bcftools mpileup -f {ref_fa_file} --regions-file {row['region_file']} --bam-list {bam_list_file} | bcftools call --ploidy 1 -vm -O z > {out_file}"
        
        return cmd
    
    df = chunk_df.copy()
    df['region_file'] = df.apply(write_region_file, axis=1)
    df['vcf_file'] = df['region_file'].apply(
        lambda f: call_res_dir / f"{f.stem}.vcf.gz"
    )
    df['call_cmd'] = df.apply(encode_cmd, axis=1)
    
    return df

chunked_call_df = get_chunked_call_df(chunk_df)

chunked_call_df

## 3. Run

### (1) call chunked vcf

cmds_to_run = [
    row['call_cmd']
    for _, row
    in chunked_call_df.iterrows()
    if not row['vcf_file'].exists()
]
len(cmds_to_run)

pool_num = 4

out, elp = execute_command(
    cmds_to_run,
    pool_num,
)

# elp
343.4221541881561

sum([v.returncode for v in out])

### (2) concat

chunk_list_file = chunked_call_dir / 'chunk_list'
chunk_list_file

with open(chunk_list_file, 'w') as f:
    for v in chunked_call_df['vcf_file'].values:
        f.write(f"{v}\n")

concated_vcf_file = chunked_call_dir / 'concated.vcf'
concated_vcf_file

cmd = f"bcftools concat --file-list {chunk_list_file} -O v > {concated_vcf_file}"
cmd

# run
out = subprocs(cmd)

elp

### (3) norm

normed_vcf_file = out_vcf_file
normed_vcf_file

cmd = f"bcftools norm -O v -f {ref_file} -d all {concated_vcf_file} > {normed_vcf_file}"
cmd

# run
out = subprocs(cmd)

elp
